JavaEnterprise

+++ src/main/java/br/com/fiap/nano/dao/FornecedorDao.java
@@ -2,5 +2,5 @@ package br.com.fiap.nano.dao;
 
 import br.com.fiap.nano.entity.Fornecedor;
 
-public interface FornecedorDao extends GenericDao<Fornecedor, Long>{
+public interface FornecedorDao extends GenericDao<Fornecedor, Integer>{
 }
