+++ src/main/java/br/com/fiap/nano/dao/ProdutoDao.java
@@ -2,5 +2,5 @@ package br.com.fiap.nano.dao;
 
 import br.com.fiap.nano.entity.Produto;
 
-public interface ProdutoDao extends GenericDao<Produto, Long> {
+public interface ProdutoDao extends GenericDao<Produto, Integer> {
 }
