+++ src/main/java/br/com/fiap/nano/dao/impl/FornecedorDaoImpl.java
@@ -5,7 +5,7 @@ import br.com.fiap.nano.entity.Fornecedor;
 
 import javax.persistence.EntityManager;
 
-public class FornecedorDaoImpl extends GenericDaoImpl<Fornecedor, Long> implements FornecedorDao {
+public class FornecedorDaoImpl extends GenericDaoImpl<Fornecedor, Integer> implements FornecedorDao {
     public FornecedorDaoImpl(EntityManager em) {
         super(em);
     }
